+++ Riot/Modules/Communities/Home/GroupHomeViewController.m
@@ -867,14 +867,13 @@
             __weak typeof(self) weakSelf = self;
             [self startActivityIndicator];
             
-            [self.mxSession.matrixRestClient roomIDForRoomAlias:roomIdOrAlias success:^(NSString *roomId) {
-                
+            [self.mxSession.matrixRestClient resolveRoomAlias:roomIdOrAlias success:^(MXRoomAliasResolution *resolution) {
                 if (roomId && weakSelf)
                 {
                     typeof(self) self = weakSelf;
                     
                     [self stopActivityIndicator];
-                    [self didSelectRoomId:roomId];
+                    [self didSelectRoomId:resolution.roomId];
                 }
                 
             } failure:^(NSError *error) {
@@ -890,7 +889,7 @@
         // Open the group or preview it
         NSString *fragment = [NSString stringWithFormat:@"/group/%@",
                         [MXTools encodeURIComponent:absoluteURLString]];
-        [[AppDelegate theDelegate] handleUniversalLinkFragment:fragment];
+        [[AppDelegate theDelegate] handleUniversalLinkFragment:fragment fromURL:URL];
     }
     
     return shouldInteractWithURL;