+++ Riot/Modules/CreateRoom/EnterNewRoomDetails/EnterNewRoomDetailsViewModel.swift
@@ -115,6 +115,7 @@ final class EnterNewRoomDetailsViewModel: EnterNewRoomDetailsViewModelType {
             fatalError("[EnterNewRoomDetailsViewModel] createRoom: room name cannot be nil.")
         }
         
+        viewState = .loading
         currentOperation = session.createRoom(
             withName: roomName,
             joinRule: roomCreationParameters.joinRule,
@@ -125,6 +126,8 @@ final class EnterNewRoomDetailsViewModel: EnterNewRoomDetailsViewModelType {
             completion: { response in
               switch response {
               case .success(let room):
+                  self.viewState = .loaded
+                  
                   if let parentSpace = self.parentSpace {
                       self.add(room, to: parentSpace)
                   } else {